+++ demisto_sdk/commands/common/hook_validations/pack_unique_files.py
@@ -25,8 +25,6 @@ from demisto_sdk.commands.common.constants import (  # PACK_METADATA_PRICE,
     PACKS_PACK_META_FILE_NAME, PACKS_README_FILE_NAME,
     PACKS_WHITELIST_FILE_NAME)
 from demisto_sdk.commands.common.errors import Errors
-from demisto_sdk.commands.common.hook_validations.author_image import \
-    AuthorImageValidator
 from demisto_sdk.commands.common.hook_validations.base_validator import \
     BaseValidator
 from demisto_sdk.commands.common.hook_validations.readme import ReadMeValidator
@@ -200,13 +198,22 @@ class PackUniqueFilesValidator(BaseValidator):
             return False
         return True
 
+    def validate_author_image_exists(self):
+        if self.metadata_content.get(PACK_METADATA_SUPPORT) == 'partner':
+            author_image_path = os.path.join(self.pack_path, 'Author_image.png')
+            if not os.path.exists(author_image_path):
+                if self._add_error(Errors.author_image_is_missing(author_image_path), file_path=author_image_path):
+                    return False
+
+        return True
+
     def validate_pack_readme_file_is_not_empty(self):
         """
         Validates that README.md file is not empty for partner packs and packs with use cases
         """
         if (self.support == 'partner' or self._contains_use_case()) and self._check_if_file_is_empty(self.readme_file):
-            self._add_error(Errors.empty_readme_error(), self.readme_file)
-            return False
+            if self._add_error(Errors.empty_readme_error(), self.readme_file):
+                return False
 
         return True
 
@@ -221,8 +228,8 @@ class PackUniqueFilesValidator(BaseValidator):
             pack_readme = self._read_file_content(self.readme_file)
             readme_content = pack_readme.lower().strip()
             if metadata_description == readme_content:
-                self._add_error(Errors.readme_equal_description_error(), self.readme_file)
-                return False
+                if self._add_error(Errors.readme_equal_description_error(), self.readme_file):
+                    return False
 
         return True
 
@@ -566,6 +573,7 @@ class PackUniqueFilesValidator(BaseValidator):
         self.validate_pack_readme_file_is_not_empty()
         self.validate_pack_readme_and_pack_description()
         self.validate_pack_readme_images()
+        self.validate_author_image_exists()
 
         # We only check pack dependencies for -g flag
         if self.validate_dependencies:
@@ -577,18 +585,7 @@ class PackUniqueFilesValidator(BaseValidator):
             if not is_valid:
                 self._add_error(error, self.pack_path)
 
-        self.validate_author_image()
-
         return self.get_errors()
-
-    def validate_author_image(self):
-        pack_meta_file_content = self._read_metadata_content()
-        support_level: str = pack_meta_file_content.get(PACK_METADATA_SUPPORT, '')
-        author_image_validator: AuthorImageValidator = AuthorImageValidator(self.pack_path, support_level,
-                                                                            ignored_errors=self.ignored_errors,
-                                                                            print_as_warnings=self.print_as_warnings,
-                                                                            suppress_print=self.suppress_print)
-        return author_image_validator.is_valid()
 
     # pack dependencies validation
     def validate_pack_dependencies(self):