+++ demisto_sdk/commands/common/tests/description_test.py
@@ -122,14 +122,15 @@ def test_demisto_in_description(repo):
 
     pack = repo.create_pack('PackName')
     integration = pack.create_integration('IntName')
-
-    description_path = glob.glob(os.path.join(os.path.dirname(integration.yml.path), '*_description.md'))[0]
-
-    with open(description_path, 'w') as f:
-        f.write('This checks if we have the word Demisto in the description.')
+    integration.create_default_integration()
+    integration.description.write('This checks if we have the word Demisto in the description.')
 
     with ChangeCWD(repo.path):
         description_validator = DescriptionValidator(integration.yml.path)
+
+        assert not description_validator.verify_demisto_in_description_content()
+
+        description_validator = DescriptionValidator(integration.description.path)
 
         assert not description_validator.verify_demisto_in_description_content()
 
@@ -148,12 +149,14 @@ def test_demisto_not_in_description(repo):
 
     pack = repo.create_pack('PackName')
     integration = pack.create_integration('IntName')
+    integration.create_default_integration()
+    integration.description.write('This checks if we have the word XSOAR in the description.')
 
-    description_path = glob.glob(os.path.join(os.path.dirname(integration.yml.path), '*_description.md'))[0]
+    with ChangeCWD(repo.path):
+        description_validator = DescriptionValidator(integration.yml.path)
 
-    with open(description_path, 'w') as f:
-        f.write('This checks if we have the word XSOAR in the description.')
+        assert description_validator.verify_demisto_in_description_content()
 
-    description_validator = DescriptionValidator(integration.yml.path)
+        description_validator = DescriptionValidator(integration.description.path)
 
-    assert description_validator.verify_demisto_in_description_content()
+        assert description_validator.verify_demisto_in_description_content()