+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 * Added to **validate** command a validation which verifies that each script that appears in incident fields, layouts or layoutscontainers exists in the id_set json.
 
+# 1.4.7
+* Fixed an issue where the **format** command has incorrectly recognized on which files to run when running using git.
+* Fixed an issue where author image validations were not checked properly.
+* Fixed an issue where new old-formatted scripts and integrations were not validated.
+* Fixed an issue where the wording in the from version validation error for subplaybooks was incorrect.
+* Fixed an issue where the **update-release-notes** command used the old docker image version instead of the new when detecting a docker change.
+* Fixed an issue where the **generate-test-playbook** command used an incorrect argument name as default
+* Fixed an issue where the **json-to-outputs** command used an incorrect argument name as default when using `-d`.
+* Fixed an issue where validations failed while trying to validate non content files.
+* Fixed an issue where README validations did not work post VS Code formatting.
+* Fixed an issue where the description validations were inconsistent when running through an integration file or a description file.
+
 # 1.4.6
 * Fixed an issue where **validate** suggests, with no reason, running **format** on missing mandatory keys in yml file.
 * Skipped existence of TestPlaybook check on community and contribution integrations.
@@ -27,6 +39,7 @@
 * Fixed an issue where the validations performed after the **format** command were inconsistent with **validate**.
 * Added to the **validate** command a validation for the author image.
 * Updated the **create-content-artifacts** command to support generic modules, definitions, fields and types.
+* Added an option to ignore errors for file paths and not only file name in .pack-ignore file.
 
 # 1.4.5
 * Enhanced the **postman-codegen** command to name all generated arguments with lower case.